requirement.txt
Flask==3.0.3
gspread==6.1.2
oauth2client==4.1.3
python-dotenv==1.0.1